+++ dev-bin/run_convergence_tests.pl
@@ -368,11 +368,13 @@ sub run_test_cases {
         my @output_history;
         for ( my $iteration = 1 ; $iteration <= $iteration_max ; $iteration++ )
         {
+	    # FIXME: remove the -irc flag below when several test cases
+	    # are corrected to avoid out-of-bounds integer flags
             my $err = Perl::Tidy::perltidy(
                 source      => \$source,
                 destination => \$output,
                 perltidyrc  => \$params,
-                argv        => '',         # don't let perltidy look at my @ARGV
+                argv        => '-irc=1',  # no warnings messages on bad integers
                 stderr    => \$stderr_string,
                 errorfile => \$errorfile_string, # not used when -se flag is set
             );