+++ lib/Perl/Tidy/Diagnostics.pm
@@ -18,7 +18,7 @@ package Perl::Tidy::Diagnostics;
 use strict;
 use warnings;
 use English qw( -no_match_vars );
-our $VERSION = '20230912.04';
+our $VERSION = '20230912.05';
 
 use constant EMPTY_STRING => q{};