+++ lib/Perl/Tidy/VerticalAligner/Alignment.pm
@@ -9,7 +9,7 @@ package Perl::Tidy::VerticalAligner::Alignment;
 use strict;
 use warnings;
 
-our $VERSION = '20230912.04';
+our $VERSION = '20230912.05';
 
 sub new {
     my ( $class, $rarg ) = @_;